Roll a small git stack
I find I apply the same operations to a list of branches all the time..
so why not make a small stack that lives in the projects file system
that can be used to apply operations to each member of the stack.

So far `rebase_each` and `push_each` are implemented.